backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.25.0
motor>=3.3.1
pymongo>=4.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic[email]
groq>=0.11.0
PyPDF2>=3.0.1
httpx>=0.28.0
bcrypt==4.0.1
passlib>=1.7.4
PyJWT>=2.11.0
python-multipart>=0.0.22
python-jose>=3.5.0
email-validator>=2.0.0
starlette>=0.37.0
pymupdf>=1.23.0
